update usage of utility
